requirements.txt
telepot
google-generativeai
bleach
python-telegram-bot~=22.0
python-telegram-bot[job-queue]
